Move Subscripte to extern component
